try to instantiate trigger on startup
